+++ src/main/java/com/dobugs/yologaauthenticationapi/auth/AuthInterceptor.java
@@ -1,4 +1,4 @@
-package com.dobugs.yologaauthenticationapi.config.auth;
+package com.dobugs.yologaauthenticationapi.auth;
 
 import java.lang.annotation.Annotation;
 
@@ -6,8 +6,8 @@ import org.springframework.web.cors.CorsUtils;
 import org.springframework.web.method.HandlerMethod;
 import org.springframework.web.servlet.HandlerInterceptor;
 
-import com.dobugs.yologaauthenticationapi.config.dto.response.ServiceToken;
-import com.dobugs.yologaauthenticationapi.config.exception.AuthorizationException;
+import com.dobugs.yologaauthenticationapi.auth.dto.response.ServiceToken;
+import com.dobugs.yologaauthenticationapi.auth.exception.AuthorizationException;
 import com.dobugs.yologaauthenticationapi.repository.TokenRepository;
 
 import jakarta.servlet.http.HttpServletRequest;
