+++ src/main/java/com/dobugs/yologaauthenticationapi/auth/TokenResolver.java
@@ -1,4 +1,4 @@
-package com.dobugs.yologaauthenticationapi.config.auth;
+package com.dobugs.yologaauthenticationapi.auth;
 
 import java.net.URLDecoder;
 import java.nio.charset.StandardCharsets;
@@ -9,7 +9,7 @@ import org.springframework.web.context.request.NativeWebRequest;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.method.support.ModelAndViewContainer;
 
-import com.dobugs.yologaauthenticationapi.config.exception.AuthorizationException;
+import com.dobugs.yologaauthenticationapi.auth.exception.AuthorizationException;
 
 import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
