+++ src/main/java/com/dobugs/yologaauthenticationapi/auth/dto/response/ServiceToken.java
@@ -1,4 +1,4 @@
-package com.dobugs.yologaauthenticationapi.config.dto.response;
+package com.dobugs.yologaauthenticationapi.auth.dto.response;
 
 public record ServiceToken(Long memberId, String provider, String tokenType, String token) {
 }
