+++ src/main/java/com/dobugs/yologaauthenticationapi/exception/ControllerAdvice.java
@@ -4,7 +4,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 
-import com.dobugs.yologaauthenticationapi.config.exception.AuthorizationException;
+import com.dobugs.yologaauthenticationapi.auth.exception.AuthorizationException;
 import com.dobugs.yologaauthenticationapi.exception.dto.response.ExceptionResponse;
 import com.dobugs.yologaauthenticationapi.support.exception.OAuthException;
