+++ src/test/java/com/dobugs/yologaauthenticationapi/config/auth/FakeController.java
@@ -5,7 +5,10 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.dobugs.yologaauthenticationapi.config.dto.response.ServiceToken;
+import com.dobugs.yologaauthenticationapi.auth.Authorized;
+import com.dobugs.yologaauthenticationapi.auth.ExtractAuthorization;
+import com.dobugs.yologaauthenticationapi.auth.ValidatedRefreshToken;
+import com.dobugs.yologaauthenticationapi.auth.dto.response.ServiceToken;
 
 @RequestMapping("/api")
 @RestController
